Create LinePlot
Just the first deposition. Still needs to add geometry nodes to allow scatter plots.  But for now, this will read any number of triplet data (x, y, z) in columns from a .csv.  No need for them to be the same length.  It will also add axes and label them.  You can scale the axes as well. 
Things to do:
1. Add scatterplot capabilities
2. Add ability to name axes easily
3. Add default material for axes (if color is not wanted).